Homeworks/hw_30/film_library/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3

Flask-Migrate~=4.0.7
alembic~=1.14.0
SQLAlchemy~=2.0.36
Flask-JWT-Extended~=4.7.1

python-dotenv~=1.0.1